+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.2] - 2018-04-11
+### Fixed
+- Incorrect processing of large files
+- Error handled trigger
+
+### Changed 
+- Skipping files larger than 70 MB (Unicheck max file upload size)
+- Event observe validation
+- Show Unicheck ID new the student file
+- Similarity score colorful background
+- Logo and link image quality
+- Rename settings
+
 ## [1.2.1] - 2017-12-14
 ### Added
 - Add unit tests
@@ -39,6 +52,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Change validation mode
 - Massive refactoring comparing to Unplag plugin
 
+[1.2.2]: https://github.com/unicheck/moodle-plagiarism_unicheckcorp/releases/tag/v1.2.2
 [1.2.1]: https://github.com/unicheck/moodle-plagiarism_unicheckcorp/releases/tag/v1.2.1
 [1.2.0]: https://github.com/unicheck/moodle-plagiarism_unicheckcorp/releases/tag/v1.2.0
 [1.1.0]: https://github.com/unicheck/moodle-plagiarism_unicheckcorp/releases/tag/v1.1.0